toolkit/cli/command/requirements.txt
typer==0.15.4
rich==14.0.0
python-dotenv==1.0.1
pydantic==2.11.5
requests==2.32.3
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1